Don't need to construct a temporary half from the half from which we'll extract a float.

+++ API-Samples/03-init_device/03-init_device.cpp
@@ -401,7 +401,7 @@ struct pixel_traits<float> {
     static float translate(const float& pixel) { return pixel; }
 
     static float translate(half pixel) {
-        return half(pixel);
+        return pixel;
     }
 
     static float translate(uchar pixel) {
